+++ scripts/mp7patch.py
@@ -5,7 +5,6 @@ import re
 
 import toolbox as tb
 
-
 def replace_area_constraints(filename):
     content = tb.read_file(filename)
 
@@ -20,10 +19,11 @@ def replace_area_constraints(filename):
     if count != 1:
         raise RuntimeError("Could not replace add_cells_to_pblock line.")
 
-    with open(filename, "wb") as fp:
+    print(content)
+
+    with open(filename, "w") as fp:
         fp.write(content)
         logging.info("Successfully patched area_constraints file '{}'".format(filename))
-
 
 def replace_brd_decl(filename):
     content = tb.read_file(filename)
@@ -39,10 +39,9 @@ def replace_brd_decl(filename):
     if count != 1:
         raise RuntimeError("Could not replace the CROSS_REGION value.")
 
-    with open(filename, "wb") as fp:
+    with open(filename, "w") as fp:
         fp.write(content)
         logging.info("Successfully patched brd_decl file '{}'".format(filename))
-
 
 def insert_l1a_ttc(filename):
     content = tb.read_file(filename)
@@ -54,29 +53,30 @@ def insert_l1a_ttc(filename):
     if count != 1:
         raise RuntimeError("Could not insert l1a ttc port.")
 
-    with open(filename, "wb") as fp:
+    with open(filename, "w") as fp:
         fp.write(content)
         logging.info("Successfully patched l1a_ttc file '{}'".format(filename))
 
+## HB 2022-09-16: with IPBB 0.5.2 'scripts/firmware/dep_tree/VivadoScriptWriter.py' does not exist anymore in environment "ipbb-0.5.2".
+## Statements "set_property ..." are now located in '../mp7_ugt_legacy/firmware/ucf/ugt_strategy.tcl'. No need for patch anymore.
 
-def append_vivado_rules(filename):
-    rules = [
-        "      ## HB 2017-04-06: inserted -assert = 1, which means \"Enable VHDL assert statements to be evaluated\"",
-        "      ## (see UG901 (v2016.3) October 21, 2016, page 12)",
-        "      write( \'set_property \"steps.synth_design.args.assert\" 1 [get_runs synth_1]\' )",
+#def append_vivado_rules(filename):
+    #rules = [
+        #"      ## HB 2017-04-06: inserted -assert = 1, which means \"Enable VHDL assert statements to be evaluated\"",
+        #"      ## (see UG901 (v2016.3) October 21, 2016, page 12)",
+        #"      write( \'set_property \"steps.synth_design.args.assert\" 1 [get_runs synth_1]\' )",
 
-        "      ## HB 2017-04-06: for timimg optimisation",
-        "      write( \'set_property strategy Performance_ExplorePostRoutePhysOpt [get_runs impl_1]\' )",
+        #"      ## HB 2017-04-06: for timimg optimisation",
+        #"      write( \'set_property strategy Performance_ExplorePostRoutePhysOpt [get_runs impl_1]\' )",
 
-        "      ## EOF",
-    ]
+        #"      ## EOF",
+    #]
 
-    with open(filename, "a") as fp:
-        for rule in rules:
-            fp.write(rule)
-            fp.write("\n")
-        logging.info("Successfully patched VivadoScriptWriter file '{}'".format(filename))
-
+    #with open(filename, "a") as fp:
+        #for rule in rules:
+            #fp.write(rule)
+            #fp.write("\n")
+        #logging.info("Successfully patched VivadoScriptWriter file '{}'".format(filename))
 
 def patch_all(projectpath):
     """Batch patch all firmware files."""
@@ -85,20 +85,18 @@ def patch_all(projectpath):
     brd_decl_path = 'boards/mp7/base_fw/mp7xe_690/firmware/hdl/mp7_brd_decl.vhd'
     area_constraints_path = 'boards/mp7/base_fw/common/firmware/ucf/area_constraints.tcl'
     mp7xe_690_path = 'boards/mp7/base_fw/mp7xe_690/firmware/hdl/mp7xe_690.vhd'
-    script_writer_path = 'scripts/firmware/dep_tree/VivadoScriptWriter.py'
+    #script_writer_path = 'scripts/firmware/dep_tree/VivadoScriptWriter.py'
 
-    replace_brd_decl(os.path.join(root_path, brd_decl_path))
+    #append_vivado_rules(os.path.join(root_path, script_writer_path))
     replace_area_constraints(os.path.join(root_path, area_constraints_path))
+    replace_brd_decl(os.path.join(root_path, brd_decl_path))
     insert_l1a_ttc(os.path.join(root_path, mp7xe_690_path))
-    append_vivado_rules(os.path.join(root_path, script_writer_path))
-
 
 def parse_args():
     """Parse command line arguments."""
     parser = argparse.ArgumentParser()
     parser.add_argument('path', type=os.path.abspath, help="path to the cactusupgrades dir of the fw tag which should be patched")
     return parser.parse_args()
-
 
 def main():
     args = parse_args()
@@ -114,7 +112,6 @@ def main():
         raise
 
     logging.info("done.")
-
 
 if __name__ == '__main__':
     main()